+++ cypress/integration/executions_spec.js
@@ -414,6 +414,7 @@ describe('Dashboard Executions Page', () => {
       cy.get('@executionGraphNodes').eq(1).should('have.text', 'HelloWorld');
 
       cy.contains('.sidebar__nav--back', 'Back to Executions').click();
+      cy.wait(1000);
       cy.get(`.td a[href="/executions/execution/${secondExecutionArn}"]`).click();
 
       cy.contains('.heading--medium', 'Visual').should('exist');
@@ -462,7 +463,7 @@ describe('Dashboard Executions Page', () => {
       cy.get('#search').as('search-input');
       cy.get('@search-input').click().type(search).type('{enter}');
 
-      cy.get('.table__main-asset > a').click();
+      cy.get('.table__main-asset > a').click({ force: true });
 
       // Breakcrumb <Link> contain correct query params
       cy.get('.breadcrumb > :nth-child(2) > a')